Skin_Cancer_Detection/src/requirements.txt
# Web Server & API
flask
flask-cors

# Deep Learning & Transformers
torch
torchvision
transformers
safetensors
accelerate

# Data Processing & Math
pandas
numpy
pillow
scikit-learn

# Utilities
huggingface-hub